Fix a Gmail search API bug
Summary:
This fixes a bug reported by one of our customers.

Sending a search request with `offset=30` and `limit=30` would return zero elements. This is because we were asking the Gmail API for at most `limit` and offsetting the results //after// receiving them. Instead, we need to ask the Gmail API for `limit + offset` and offset them.

Test Plan: Tested manually. Checked that a query with `?limit=30&offset=30` returned more than zero elements.

Reviewers: spang

Reviewed By: spang

Subscribers: jenkins

Differential Revision: https://phab.nylas.com/D2960